Replace f32 by f64

+++ src/configs/raw.rs
@@ -142,7 +142,7 @@ pub mod routing {
     #[derive(Debug, Deserialize)]
     pub struct Entry {
         pub id: SimpleId,
-        pub alpha: Option<f32>,
+        pub alpha: Option<f64>,
     }
 }
